+++ packages/xo-web/src/common/xo/index.js
@@ -1973,8 +1973,8 @@ export const getRemote = remote =>
   )
 
 export const createRemote = (name, url, options) =>
-  _call('remote.create', { name, url, options })::tap(
-    subscribeRemotes.forceRefresh
+  _call('remote.create', { name, url, options })::tap(remote =>
+    testRemote(remote).catch(noop)
   )
 
 export const deleteRemote = remote =>
@@ -1997,8 +1997,8 @@ export const deleteRemotes = remotes =>
   )
 
 export const enableRemote = remote =>
-  _call('remote.set', { id: resolveId(remote), enabled: true })::tap(
-    subscribeRemotes.forceRefresh
+  _call('remote.set', { id: resolveId(remote), enabled: true })::tap(() =>
+    testRemote(remote).catch(noop)
   )
 
 export const disableRemote = remote =>
@@ -2007,8 +2007,8 @@ export const disableRemote = remote =>
   )
 
 export const editRemote = (remote, { name, url, options }) =>
-  _call('remote.set', resolveIds({ remote, name, url, options }))::tap(
-    subscribeRemotes.forceRefresh
+  _call('remote.set', resolveIds({ remote, name, url, options }))::tap(() =>
+    testRemote(remote).catch(noop)
   )
 
 export const listRemote = remote =>
@@ -2023,9 +2023,9 @@ export const listRemoteBackups = remote =>
   )
 
 export const testRemote = remote =>
-  _call('remote.test', resolveIds({ id: remote }))::tap(null, err =>
-    error(_('testRemote'), err.message || String(err))
-  )
+  _call('remote.test', resolveIds({ id: remote }))
+    ::tap(null, err => error(_('testRemote'), err.message || String(err)))
+    ::pFinally(subscribeRemotes.forceRefresh)
 
 // File restore  ----------------------------------------------------